Add support for myriad to XP calculation
Complicated because of things like Adaptable (free level 0 swap means 3
free swaps) AND Arcane Research.
The wording of the rule also seems to suggest that we might someday see
cards with the same name (i.e. seeker upgrades) that could be Myriad and
share the deckbuilding rule of only paying once, which is why I have
code based on real_name in there